gmail: Create newJWTClient() function.
This function is an alternate code path for
authentication.

The main purpose of this function is to be able
to authenticate using a JSON Web Token (JWT)
against Gmail with room for impersonation of
any user of the domain.

The next patch will wire up this to the main()
arguments parsing code.