gmail: Use Label Get() in order to get better progress estimate.
The ResultSizeEstimate field is only an estimate.

This leads the backup progress computation to use a wrong total
account of mails hence displaying bogus results in non small
mailboxes.

This commit creates a method to be used to compute something
nearing the mail count (the label count) in order to compute
a better backup progress estimation.

This works by retrieving the Labels() then iterating on the
resulting list and filtering only by the label names that
make sense.

The two cases this patch implements are:

1 ) No label is to be used for filtering:

Then we take into labels in the { "DRAFT" , "INBOX", "SENT", "TRASH" }
set. The rationale for this is that these labels fit 'real'
mails contrary to the other labels.

Note that the labeled "TRASH" mail account must be substracted
from the account computed above.

2 ) A label was specified in order to make a selective backup:

Then simply account the number of mails having this label.

Finally the messages sent to feed the progress display goroutine
takes into account the relevant label count of each mails if we
are into the 1) case.
Again relevant labels are in { "DRAFT" , "INBOX", "SENT", "TRASH" }.